Add search bar below navigation bar
Search is fully funtioning, but is limited to exact beginning matches of the product name.

For instance, product name = myname is BOB
The search will have to be = myna or myname is
Searching for MyName or is BOB will not return anything.
This could potentially be fixed with Firebase Functions that was released yesterday…